Implemented call to CcdApi and enabled func test
