+++ src/integrationTest/resources/ccd/response/search-result-empty.json
@@ -1,0 +1,4 @@
+{
+  "total": 0,
+  "cases": []
+}
